Re-organise the code
Move all the OpenAPI-specific things into their own folder/namespace.
