+++ src/TodoApp/Models/CreateTodoItemModel.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Martin Costello, 2024. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
+using TodoApp.OpenApi;
+
 namespace TodoApp.Models;
 
 /// <summary>
