+++ src/TodoApp/OpenApi/AspNetCore/AspNetCoreOpenApiEndpoints.cs
@@ -2,9 +2,8 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
 using Microsoft.OpenApi.Models;
-using TodoApp.OpenApi;
 
-namespace TodoApp;
+namespace TodoApp.OpenApi.AspNetCore;
 
 public static class AspNetCoreOpenApiEndpoints
 {
