+++ src/TodoApp/OpenApi/IOpenApiExampleMetadata.cs
@@ -4,7 +4,7 @@
 using System.Text.Json.Serialization;
 using Microsoft.OpenApi.Any;
 
-namespace TodoApp;
+namespace TodoApp.OpenApi;
 
 /// <summary>
 /// Defines an OpenAPI example.
