+++ src/TodoApp/OpenApi/OpenApiGuidExampleAttribute.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Martin Costello, 2024. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-namespace TodoApp;
+namespace TodoApp.OpenApi;
 
 /// <summary>
 /// An attribute representing an example for an OpenAPI operation parameter.
